Add SwitcherProvider
