+++ package.json
@@ -52,6 +52,7 @@
     "babel-cli": "^6.24.1",
     "babel-eslint": "^7.2.1",
     "babel-preset-es2015": "^6.24.1",
+    "babel-preset-es2015-rollup": "^3.0.0",
     "babel-preset-react": "^6.24.1",
     "babel-preset-stage-0": "^6.24.1",
     "babel-register": "^6.14.0",
