+++ rollup.config.js
@@ -11,10 +11,11 @@ export default {
     babel({
       exclude: 'node_modules/**',
       babelrc: false,
-      presets: [['es2015', { modules: false }], 'stage-0', 'react']
+      presets: ['es2015-rollup', 'stage-0', 'react']
     })
   ],
   external: ['react', 'prop-types'],
+  exports: 'named',
   globals: {
     react: 'React',
     'prop-types': 'PropTypes'
