+++ src/helpers.js
@@ -1,5 +1,14 @@
 import { Children } from 'react';
 
+// http://stackoverflow.com/a/2117523
+export function generateGuid() {
+  return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, function(c) {
+    const r = (Math.random() * 16) | 0;
+    const v = c === 'x' ? r : (r & 0x3) | 0x8;
+    return v.toString(16);
+  });
+}
+
 export function currentPath(location) {
   const path = decodeURI(window.location[location].slice(1).split('?')[0]);
   if (path.charAt(0) !== '/') {
